# api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LongLink Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/computes": {
      "post": {
        "summary": "Create Compute Registry",
        "description": "Register a compute target and queue its initial creation.",
        "operationId": "create_compute_registry_api_v1_computes_post",
        "parameters": [
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputeRegistryCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeRegistryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Compute Registries",
        "description": "Return all registered compute backends.",
        "operationId": "list_compute_registries_api_v1_computes_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ComputeRegistryResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/computes/{registry_id}": {
      "get": {
        "summary": "Get Compute Registry",
        "description": "Return one compute backend registration.",
        "operationId": "get_compute_registry_api_v1_computes__registry_id__get",
        "parameters": [
          {
            "name": "registry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Registry Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeRegistryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Compute Registry",
        "description": "Remove one unused compute registration without changing its cluster.",
        "operationId": "delete_compute_registry_api_v1_computes__registry_id__delete",
        "parameters": [
          {
            "name": "registry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Registry Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/databases": {
      "post": {
        "summary": "Create Database Registry",
        "description": "Register one database backend.",
        "operationId": "create_database_registry_api_v1_databases_post",
        "parameters": [
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabaseRegistryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseRegistryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Database Registries",
        "description": "Return all registered database backends.",
        "operationId": "list_database_registries_api_v1_databases_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DatabaseRegistryResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/databases/{registry_id}": {
      "get": {
        "summary": "Get Database Registry",
        "description": "Return one database backend registration.",
        "operationId": "get_database_registry_api_v1_databases__registry_id__get",
        "parameters": [
          {
            "name": "registry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Registry Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseRegistryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Database Registry",
        "description": "Delete one unused database backend registration.",
        "operationId": "delete_database_registry_api_v1_databases__registry_id__delete",
        "parameters": [
          {
            "name": "registry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Registry Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/databases/{registry_id}/usage": {
      "get": {
        "summary": "Get Database Usage",
        "description": "Query point-in-time storage usage from the live database backend, not persisted desired state.\n\nThe result is diagnostic and depends on backend availability.",
        "operationId": "get_database_usage_api_v1_databases__registry_id__usage_get",
        "parameters": [
          {
            "name": "registry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Registry Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Get Database Usage Api V1 Databases  Registry Id  Usage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComputeRegistryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "kubeconfig": {
            "additionalProperties": true,
            "type": "object",
            "title": "Kubeconfig"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kubeconfig"
        ],
        "title": "ComputeRegistryCreate",
        "description": "Validate one compute registry creation payload."
      },
      "ComputeRegistryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gateway_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway Url"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "gateway_url",
          "status"
        ],
        "title": "ComputeRegistryResponse",
        "description": "Describe one compute backend without exposing its private connection state or secrets."
      },
      "DatabaseRegistryCreate": {
        "properties": {
          "host": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port"
          },
          "sslmode": {
            "$ref": "#/components/schemas/DatabaseSSLMode",
            "default": "require"
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Password"
          },
          "username": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Username"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "host",
          "port",
          "password",
          "username",
          "name"
        ],
        "title": "DatabaseRegistryCreate",
        "description": "Validate one database registry creation payload."
      },
      "DatabaseRegistryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "host": {
            "type": "string",
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "sslmode": {
            "$ref": "#/components/schemas/DatabaseSSLMode"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "host",
          "port",
          "sslmode",
          "username"
        ],
        "title": "DatabaseRegistryResponse",
        "description": "Describe one database backend while filtering its administrator password.\n\nNon-secret connection metadata remains available for administrator diagnostics."
      },
      "DatabaseSSLMode": {
        "type": "string",
        "enum": [
          "disable",
          "allow",
          "prefer",
          "require",
          "verify-ca",
          "verify-full"
        ],
        "title": "DatabaseSSLMode",
        "description": "Supported PostgreSQL SSL modes."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_ComputeRegistryResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ComputeRegistryResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[ComputeRegistryResponse]"
      },
      "Page_DatabaseRegistryResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatabaseRegistryResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[DatabaseRegistryResponse]"
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "creating",
          "running"
        ],
        "title": "Status",
        "description": "Lifecycle states shared by Platform-managed resources."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}